# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Enlace Group Loan API",
    "description": "API for group lending: onboarding, loans, repayments, trust and credibility.\n\nDocumentation layout: **Admin** and **Users** are top-level groups in the sidebar when your viewer supports `x-tagGroups` (e.g. Redoc, Stoplight). Sub-bullets (Dashboard, KYC, …) are OpenAPI tags. **Note:** borrower KYC lives under `/auth/kyc`, `/auth/nin`, `/auth/lookup` (shown as **Users · KYC**). Swagger UI often lists tags flat and alphabetically.",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "API root"
    }
  ],
...
  "paths": {
...
    "/dashboard": {
      "get": {
        "tags": [
          "Users · Dashboard"
        ],
        "summary": "Get dashboard",
        "description": "Returns the full dashboard for the authenticated user. `userType` is `\"group\"` if the user is an active member of a group, or `\"individual\"` for standalone users. The `view` field is `\"onboarding\"` until the user (or all group members) complete KYC; then `\"full\"`. Fields marked as group-only are null for individuals; fields marked individual-only are null for group members.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardData"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/admin/dashboard": {
      "get": {
        "tags": [
          "Admin · Dashboard"
        ],
        "summary": "Get admin dashboard",
        "description": "Returns KPIs (portfolio value, active loans, default rate, pending KYC), portfolio growth by week, system alerts, recent loan operations, and distribution by loan type.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "kpis": {
                      "type": "object",
                      "properties": {
                        "portfolioValue": {
                          "type": "number"
                        },
                        "portfolioValueChangePercent": {
                          "type": "number"
                        },
                        "activeLoans": {
                          "type": "integer"
                        },
                        "activeLoansChangePercent": {
                          "type": "number"
                        },
                        "defaultRate": {
                          "type": "number"
                        },
                        "defaultRateChangePercent": {
                          "type": "number"
                        },
                        "pendingKyc": {
                          "type": "integer"
                        },
                        "pendingKycChangePercent": {
                          "type": "number"
                        }
                      }
                    },
                    "portfolioGrowth": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "weekLabel": {
                            "type": "string"
                          },
                          "total": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "alerts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "severity": {
                            "type": "string",
                            "enum": [
                              "error",
                              "warning",
                              "info"
                            ]
                          }
                        }
                      }
                    },
                    "recentLoans": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loanId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "borrowerName": {
                            "type": "string"
                          },
                          "loanType": {
                            "type": "string",
                            "nullable": true
                          },
                          "amount": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          },
                          "currency": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "distributionByType": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string"
                          },
                          "percent": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/groups": {
      "get": {
        "tags": [
          "Admin · Users & groups"
        ],
        "summary": "List credit groups (admin)",
        "description": "Paginated directory with initiator, member count, credit limit fields, and derived onboardingStatus for the admin Credit Groups table.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onboardingStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "PENDING_KYC",
                "ONBOARDING",
                "FLAGGED"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "groups": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access denied"
          }
        }
      },
      "post": {
        "tags": [
          "Admin · Users & groups"
        ],
        "summary": "Create credit group (admin)",
        "description": "Same payload as `POST /groups` plus `createdByUserId` (existing user who becomes creator). Sends invites when `members` provided.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "states",
                  "expectedLoan",
                  "createdByUserId"
                ],
                "properties": {
                  "createdByUserId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "states": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "expectedLoan": {
                    "type": "number",
                    "minimum": 0
                  },
                  "targetCredit": {
                    "type": "number"
                  },
                  "minimumAmount": {
                    "type": "number"
                  },
                  "maximumAmount": {
                    "type": "number"
                  },
                  "members": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "fullName",
                        "email"
                      ],
                      "properties": {
                        "fullName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "phone": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation or business rule"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access denied"
          }
        }
      }
    },
...
    "/groups": {
      "post": {
        "tags": [
          "Users · Groups"
        ],
        "summary": "Create group",
        "description": "Creates a group with the caller as creator. Optionally include `members` to add members and send them invite emails (existing users get a group_member invite; new users get a pending group_invite until they sign up).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "states",
                  "expectedLoan"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1
                  },
                  "states": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "example": [
                      "Lagos",
                      "Oyo"
                    ]
                  },
                  "expectedLoan": {
                    "type": "number",
                    "minimum": 0
                  },
                  "targetCredit": {
                    "type": "number",
                    "minimum": 0
                  },
                  "minimumAmount": {
                    "type": "number",
                    "minimum": 0
                  },
                  "maximumAmount": {
                    "type": "number",
                    "minimum": 0
                  },
                  "repaymentPeriod": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "repaymentType": {
                    "type": "string",
                    "enum": [
                      "daily",
                      "weekly",
                      "monthly",
                      "quarterly",
                      "yearly"
                    ]
                  },
                  "interestType": {
                    "type": "string",
                    "enum": [
                      "flat",
                      "reducingBalance"
                    ]
                  },
                  "interest": {
                    "type": "number",
                    "minimum": 0
                  },
                  "penalCharges": {
                    "type": "number",
                    "minimum": 0
                  },
                  "gracePeriod": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "gracePeriodType": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "overGracePenalCharges": {
                    "type": "number",
                    "minimum": 0
                  },
                  "ageRange": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "18-35",
                      "30-45"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE",
                      "PENDING"
                    ]
                  },
                  "members": {
                    "type": "array",
                    "description": "Optional list of members to add and send invite emails. Existing users are added as invited members; new users get a pending invite until signup.",
                    "items": {
                      "type": "object",
                      "required": [
                        "fullName",
                        "email"
                      ],
                      "properties": {
                        "fullName": {
                          "type": "string",
                          "minLength": 1,
                          "description": "Display name"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "phone": {
                          "type": "string",
                          "description": "Optional phone number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invitee already in a group / already invited"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/history": {
      "post": {
        "operationId": "getHistory",
        "tags": [
          "Users · History"
        ],
        "summary": "Get my activity timeline",
        "description": "Unified activity feed for the authenticated user, merged and sorted newest-first across:\n\n- **Loan requests** — every loan you've ever requested (group or individual)\n- **Approved** — your loans that reached APPROVED, REVIEWING, PROCESSING, DISBURSED, ACTIVE, or REPAID\n- **Repayments** — every paid repayment you've made, plus a separate `loan_repaid` event when a loan reaches REPAID\n- **Group** — other members joining or exiting your active groups (only events after you joined)\n- **Milestones** — group tier upgrades (currently a snapshot of each group's non-STANDARD `credibilityLevel`; will become a true audit log later)\n\n**Body fields**\n- `category` — one of `all`, `loan_requests`, `approved`, `repayments`, `group`, `milestones`. Defaults to `all`.\n- `startDate` / `endDate` — optional ISO datetimes filtering on each item's `at` timestamp.\n- `limit` — page size (default 20, max 100).\n- `offset` — pagination offset (default 0).\n\nEach item has a stable `id` of the form `<type>:<sourceId>` so the frontend can de-duplicate or animate transitions safely.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryRequest"
              },
              "examples": {
                "all": {
                  "summary": "All activity (default)",
                  "value": { "category": "all", "limit": 20, "offset": 0 }
                },
                "repaymentsOnly": {
                  "summary": "Repayments tab",
                  "value": { "category": "repayments", "limit": 20, "offset": 0 }
                },
                "groupTab": {
                  "summary": "Group tab",
                  "value": { "category": "group", "limit": 20, "offset": 0 }
                },
                "byDate": {
                  "summary": "Date-bounded",
                  "value": {
                    "category": "all",
                    "startDate": "2026-01-01T00:00:00.000Z",
                    "endDate": "2026-05-31T23:59:59.999Z",
                    "limit": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paginated activity timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation (e.g. invalid category, or startDate after endDate)"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/repayments": {
      "post": {
        "tags": [
          "Users · Repayments"
        ],
        "summary": "Record repayment (borrower only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "loanId",
                  "amount"
                ],
                "properties": {
                  "loanId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "amount": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount or loan not repayable"
          },
          "403": {
            "description": "Only borrower can record"
          },
          "404": {
            "description": "Loan not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from login, verify-email (first success), or signup flows that return a token"
      }
    },
    "schemas": {
...
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "targetCredit": {
            "type": "number"
          },
          "currentCreditPool": {
            "type": "number"
          },
          "credibilityScore": {
            "type": "number"
          },
          "credibilityLevel": {
            "type": "string",
            "enum": [
              "STANDARD",
              "VERIFIED_TRUST_GROUP"
            ]
          },
          "quarterlyStartDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "quarterlyEndDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "minimumAmount": {
            "type": "number",
            "nullable": true
          },
          "maximumAmount": {
            "type": "number",
            "nullable": true
          },
          "repaymentPeriod": {
            "type": "integer",
            "nullable": true
          },
          "repaymentType": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "quarterly",
              "yearly"
            ],
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "interestType": {
            "type": "string",
            "enum": [
              "flat",
              "reducingBalance"
            ],
            "nullable": true
          },
          "interest": {
            "type": "number",
            "nullable": true
          },
          "penalCharges": {
            "type": "number",
            "nullable": true
          },
          "gracePeriod": {
            "type": "integer",
            "nullable": true
          },
          "gracePeriodType": {
            "type": "string",
            "nullable": true
          },
          "overGracePenalCharges": {
            "type": "number",
            "nullable": true
          },
          "ageRange": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "states": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expectedLoan": {
            "type": "number",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PENDING"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "runningMandate": {
            "type": "boolean",
            "description": "On profile nested group: true if this user has an ACTIVE direct-debit mandate for this group within the last 12 months"
          }
        }
      },
      "HistoryRequest": {
        "type": "object",
        "description": "Body for POST /history. All fields optional; defaults to category=all, limit=20, offset=0.",
        "properties": {
          "category": {
            "type": "string",
            "enum": ["all", "loan_requests", "approved", "repayments", "group", "milestones"],
            "default": "all"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 20
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" }
        }
      },
...
      "DashboardData": {
        "type": "object",
        "description": "Unified dashboard response for both group members and individual users.",
        "properties": {
          "view": {
            "type": "string",
            "enum": ["onboarding", "full"],
            "description": "onboarding = KYC not yet complete; full = ready to transact."
          },
          "userType": {
            "type": "string",
            "enum": ["group", "individual"],
            "description": "group = active member of a credit group; individual = standalone user."
          },
          "user": {
            "type": "object",
            "properties": {
              "fullName": { "type": "string" },
              "memberIdDisplay": { "type": "string" },
              "trustLevel": { "type": "string", "enum": ["BRONZE", "SILVER", "GOLD"] },
              "currentBalance": { "type": "number" }
            }
          },
          "groupHealthPercent": {
            "type": "integer",
            "description": "Group members: % of members with KYC+income+bank done. Individual: 100 if KYC approved, proportional otherwise."
          },
          "allKycComplete": { "type": "boolean" },
          "groupOnboarding": {
            "nullable": true,
            "description": "Group member onboarding checklist. Null for individual users.",
            "type": "object",
            "properties": {
              "groupId": { "type": "string", "format": "uuid" },
              "groupName": { "type": "string" },
              "onboardingPercent": { "type": "integer" },
              "groupGoal": { "type": "number" },
              "unlockedSoFar": { "type": "number" },
              "membersReady": { "type": "integer" },
              "membersTotal": { "type": "integer" },
              "creditStatus": { "type": "string", "enum": ["locked", "unlocked"] },
              "lockReason": { "type": "string" },
              "members": { "type": "array", "items": { "type": "object" } }
            }
          },
          "individualOnboarding": {
            "nullable": true,
            "description": "Individual user onboarding checklist. Null for group members.",
            "allOf": [{ "$ref": "#/components/schemas/IndividualOnboarding" }]
          },
          "availableCreditPool": { "type": "number", "description": "Admin-approved access cap for the user's primary group (group.maximumAmount, falling back to targetCredit then currentCreditPool when unset). This is also the maximum a member can request via POST /loans/group. 0 for individual users." },
          "creditPoolUtilizationPercent": { "type": "integer" },
          "personalLimit": { "type": "number", "description": "User's individual credit limit (3× monthly income)." },
          "projectedGroupLimit": { "type": "number", "description": "Group target credit (0 for individual users)." },
          "trustBadge": { "type": "string", "description": "Display label for the user's trust level badge." },
          "badgeProgress": {
            "nullable": true,
            "type": "object",
            "properties": {
              "current": { "type": "number" },
              "target": { "type": "number" },
              "nextMilestone": { "type": "string" }
            }
          },
          "activeLoans": { "type": "integer" },
          "pendingApprovalsCount": { "type": "integer" },
          "pendingPeerApprovals": { "type": "array", "items": { "type": "object" } },
          "recentActivity": { "type": "array", "items": { "type": "object" } },
          "earnBonusPoints": {
            "type": "object",
            "properties": {
              "description": { "type": "string" },
              "invitePoints": { "type": "integer" }
            }
          },
          "groupPoolStatus": {
            "type": "array",
            "description": "One entry per group the user is active in (empty for individual users).",
            "items": {
              "type": "object",
              "properties": {
                "groupId": { "type": "string", "format": "uuid" },
                "groupName": { "type": "string" },
                "currentPool": { "type": "number" }
              }
            }
          },
          "credibilityScore": { "type": "number", "nullable": true, "description": "Highest credibility score across user's groups. Null for individual users." },
          "creditEligibility": { "type": "number" },
          "deadline": { "type": "string", "format": "date-time", "nullable": true, "description": "Group quarterly end date. Null for individual users." },
          "projectedAmount": { "type": "number" },
          "nonAcceptedMemberships": {
            "type": "array",
            "description": "Group invitees who have not yet accepted. Empty for individual users.",
            "items": {
              "type": "object",
              "properties": {
                "userId": { "type": "string", "format": "uuid", "nullable": true, "description": "Set when invitee has an account (group_members); null for email-only group_invites" },
                "inviteId": { "type": "string", "format": "uuid", "description": "Present for pending email-only invites (group_invites id); use with POST /groups/{id}/invites/{inviteId}/poke" },
                "poke": { "type": "boolean", "description": "Only on email-only invites: true if the creator already used the one-time poke" },
                "fullName": { "type": "string" },
                "email": { "type": "string" },
                "status": { "type": "string", "enum": ["INVITED"] },
                "invitedAt": { "type": "string", "format": "date-time", "nullable": true }
              }
            }
          },
          "individualMandate": {
            "nullable": true,
            "description": "Active individual mandate period info. Null for group members.",
            "allOf": [{ "$ref": "#/components/schemas/IndividualMandateInfo" }]
          }
        }
      },
...
    }
  }
}